+++ src/transport/test_transport_api_reliability.c
@@ -228,10 +228,10 @@ notify_receive (void *cls,
   {
     GNUNET_log (GNUNET_ERROR_TYPE_ERROR,
                 "Expected message %u of size %u, got %u bytes of message %u\n",
-                ntohl (hdr->num),
+                (uint32_t) ntohl (hdr->num),
                 s,
                 ntohs (hdr->header.size),
-                ntohl (hdr->num));
+                (uint32_t) ntohl (hdr->num));
     ccc->global_ret = GNUNET_SYSERR;
     GNUNET_SCHEDULER_shutdown ();
     return;
@@ -247,7 +247,7 @@ notify_receive (void *cls,
   {
     GNUNET_log (GNUNET_ERROR_TYPE_ERROR,
                 "Expected message %u with bits %u, but body did not match\n",
-                ntohl (hdr->num),
+                (uint32_t) ntohl (hdr->num),
                 (unsigned char) ntohl (hdr->num));
     ccc->global_ret = GNUNET_SYSERR;
     GNUNET_SCHEDULER_shutdown ();
@@ -258,7 +258,7 @@ notify_receive (void *cls,
   {
     GNUNET_log (GNUNET_ERROR_TYPE_DEBUG,
                 "Got message %u of size %u\n",
-                ntohl (hdr->num),
+                (uint32_t) ntohl (hdr->num),
                 ntohs (hdr->header.size));
   }
 #endif
@@ -267,7 +267,7 @@ notify_receive (void *cls,
   {
       GNUNET_log (GNUNET_ERROR_TYPE_ERROR,
                   "Message id %u is bigger than maxmimum number of messages %u expected\n",
-                  ntohl (hdr->num),
+                  (uint32_t) ntohl (hdr->num),
                   TOTAL_MSGS / xhdr);
   }
   if (0 == (n % (TOTAL_MSGS / xhdr / 100)))
